+++ Docs/Changelog.md
@@ -1,10 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [0.13.1] - 2017-12-06
+## [0.14.0] - 2017-12-06
 
 ## Changed
 - Opcode debug output now shows symbol and function names
+- Logical 'not' operation now has same precedence as 'or/and'
 
 ## Fixed
 - Parser now rejects function parameter names that collide with properties